+++ spec/files/worker_init/first_file.rb
@@ -3,7 +3,7 @@ class FirstFile
     # do nothing in this example
   end
 
-  def run(*args)
+  def run(a_file, *args)
     puts args
 
     args