+++ CHANGELOG.md
@@ -4,6 +4,7 @@ OpenStudio Analysis Gem Change Log
 Version 0.3.7 (Unreleased)
 -------------
 * Worker initilization and finalization scripts
+* Do not allow the file to proccess if the Measure Display Names are not unique
 
 Version 0.3.6
 -------------